+++ packages/core/src/SearchInput.js
@@ -4,6 +4,7 @@ import classNames from 'classnames';
 
 import icBEM from './utils/icBEM';
 import prefixClass from './utils/prefixClass';
+import getRemainingProps from './utils/getRemainingProps';
 import rowComp from './mixins/rowComp';
 import './styles/SearchInput.scss';
 
@@ -146,8 +147,10 @@ class SearchInput extends Component {
         const isLoading = this.context.status === STATUS_CODE.LOADING;
         const rootClassName = classNames(className, `${BEM.root}`);
 
+        const wrapperProps = getRemainingProps(this.props, SearchInput.propTypes);
+
         return (
-            <div className={rootClassName}>
+            <div className={rootClassName} {...wrapperProps}>
                 <div className={BEM.inputWrapper}>
                     <Icon type="search" />