+++ packages/core/src/Tag.js
@@ -8,11 +8,11 @@ import prefixClass from './utils/prefixClass';
 const COMPONENT_NAME = prefixClass('tag');
 const ROOT_BEM = icBEM(COMPONENT_NAME);
 
-function Tag({ className, children }) {
+function Tag({ className, children, ...wrapperProps }) {
     const rootClass = classNames(`${ROOT_BEM.toString()}`, className);
 
     return (
-        <span className={rootClass}>
+        <span className={rootClass} {...wrapperProps}>
             <span>{children}</span>
         </span>
     );