+++ packages/core/src/__tests__/Popup.test.js
@@ -31,11 +31,49 @@ describe('Pure <Popup>', () => {
         expect(wrapper.find(Overlay).exists()).toBeTruthy();
     });
 
+    it('render with custom node of customMessageNode', () => {
+        const customMessageNode = <img src="test-img" alt="test-img" />;
+        const wrapper = shallow(<PurePopup customMessageNode={customMessageNode} />);
+        expect(wrapper.find(customMessageNode).exists()).toBeFalsy();
+        expect(
+            wrapper.find('img[src="test-img"]').exists()
+        ).toBeTruthy();
+    });
+
+    it('renders a string messageTitle and messageDesc with <PopupMessage>', () => {
+        const customBottomArea = <img src="test-bottom-area" alt="test-img" />;
+        const wrapper = shallow((
+            <PurePopup
+                messageTitle="foo"
+                messageDesc="bar"
+                messageBottomArea={customBottomArea}
+            />
+        ));
+
+        expect(wrapper.find(PopupMessage).exists()).toBeTruthy();
+        expect(wrapper.find(PopupMessage).prop('title')).toBe('foo');
+        expect(wrapper.find(PopupMessage).prop('desc')).toBe('bar');
+        expect(wrapper.find(PopupMessage).prop('bottomArea')).toBe(customBottomArea);
+    });
+
     it('renders a string message with <PopupMessage>', () => {
         const wrapper = shallow(<PurePopup message="foo" />);
 
         expect(wrapper.find(PopupMessage).exists()).toBeTruthy();
-        expect(wrapper.find(PopupMessage).prop('text')).toBe('foo');
+        expect(wrapper.find(PopupMessage).prop('desc')).toBe('foo');
+    });
+
+    it('renders popup with large class name', () => {
+        const wrapper = shallow(<PurePopup large messageDesc="foo" />);
+
+        expect(wrapper.find('.gyp-popup--large').exists()).toBeTruthy();
+    });
+
+    it('renders popup without large class name', () => {
+        const wrapper = shallow(<PurePopup messageDesc="foo" />);
+
+        expect(wrapper.find('.gyp-popup__body--large').exists()).toBeFalsy();
+        expect(wrapper.find('.gyp-popup__container--large').exists()).toBeFalsy();
     });
 
     it('takes a valid element for message prop', () => {
@@ -102,8 +140,23 @@ describe('<PopupIcon>', () => {
 });
 
 describe('<PopupMessage>', () => {
-    it('returns a pre-configured <TextLabel> of given text', () => {
-        const wrapper = shallow(<PopupMessage text="foo" />);
-        expect(wrapper.matchesElement(<TextLabel align="center" basic="foo" />)).toBeTruthy();
+    it('returns a pre-configured layout of given props', () => {
+        const wrapper = shallow(<PopupMessage title="foo" desc="bar" bottomArea="bottomArea" />);
+        expect(wrapper.matchesElement(
+            <div>
+                <span>foo</span>
+                <span>bar</span>
+                bottomArea
+            </div>
+        )).toBeTruthy();
+    });
+
+    it('only render with desc prop', () => {
+        const wrapper = shallow(<PopupMessage desc="bar" />);
+        expect(wrapper.matchesElement(
+            <div>
+                <span>bar</span>
+            </div>
+        )).toBeTruthy();
     });
 });